Acess InfluxDB via URL instead of hostname and port
I have added a new configuration item called "url" that allows to directly configure a URL for InfluxDB. This enables accessing InfluxDB e.g. via SSL or via a reverse proxy.

Some examples for possible URL formats:

http://127.0.0.1:8086
https://127.0.1
https://1270.0.1/some/sub/path

If URL is not specified, hostname and port continue to be used as previously.